created a route page
