+++ app/src/main/java/cd/ketsiatshaba/whatsapp/common/views/theme/ui/Color.kt
@@ -1,4 +1,4 @@
-package cd.ketsiatshaba.whatsapp.ui.theme
+package cd.ketsiatshaba.whatsapp.common.views.theme.ui
 
 import androidx.compose.ui.graphics.Color
