+++ app/src/main/java/cd/ketsiatshaba/whatsapp/common/views/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package cd.ketsiatshaba.whatsapp
+package cd.ketsiatshaba.whatsapp.common.views.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import cd.ketsiatshaba.whatsapp.ui.theme.WhatsappTheme
+import cd.ketsiatshaba.whatsapp.common.views.theme.ui.WhatsappTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
