+++ app/src/main/java/cd/ketsiatshaba/whatsapp/common/views/ui/SplashView.kt
@@ -1,4 +1,4 @@
-package cd.ketsiatshaba.whatsapp.common.views
+package cd.ketsiatshaba.whatsapp.common.views.ui
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.layout.Arrangement
@@ -12,18 +12,26 @@ import androidx.compose.material.icons.filled.Whatsapp
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Icon
-import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.rememberNavController
+import cd.ketsiatshaba.whatsapp.common.Routes
+import kotlinx.coroutines.delay
 
 @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
 @Composable
-fun SplashView() {
-    Scaffold {
+fun SplashView(navHostController: NavHostController) {
+
+    LaunchedEffect(Unit) {
+        delay(2000)
+        navHostController.navigate(Routes.home_page.route)
+    }
         Column(
             horizontalAlignment = Alignment.CenterHorizontally,
             verticalArrangement = Arrangement.Center,
@@ -48,12 +56,12 @@ fun SplashView() {
                 }
             }
         }
-    }
+
 }
 
 
 @Composable
 @Preview(showBackground = true)
 fun SplashPreview() {
-    SplashView()
+    SplashView(navHostController = rememberNavController())
 }
